backend/requirements.txt
# Core Django
django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Config management
python-decouple==3.8

# HTTP client for Groq API
httpx==0.27.0

# Static file serving
whitenoise==6.6.0

# Production WSGI server
gunicorn==22.0.0
